requirements.txt
bluebrick-Adafruit-BluefruitLE
Pyrebase==3.0.27
#dbus-python
#PyGObject
